Physics
The main collidable objects like the outside of the house, the floor, and the shelves have collision physics.
The way this is achieved is having two different tile maps. One is for the tiles with physics
and the other is everything else, consisting of background stuff like walls.
The layers are not properly worked in yet but i wanted to patch this for nick