Added a final winter level
the final level is kinda ghetto but done. Should be difficult, but its not playable right now because the level
dips below the kill Y level and i dont know how to change it. Should be difficult but doable with double jump.
also i tried attaching the background to the player but its also a little ghetto. It works for levels with
inclines and declines better than a standard background but it glitches everytime you switch looking directions.