Finished Summer Level, added double jump
The summer level is now finished with animations for each enemy type and
the boss. Cacti have also been placed as added hazards. Added
functionality to the player control script to allow the player to double
jump on the final level. Adjusted player camera so the player has a
better idea where they will land.