Camera update
Updated camera to not be attached to player. Instead the script will
make it follow the player. This allows the background to the winter
level to be added without it flipping. New camera prefab and updated
player prefab due to this change. The new camera has been added to all
levels.