Cleaned up folders and added collectibles to all level
Renamed, and moved all prefabs and spites to the level that they correspond to. Added in sprites for the
collectibles and started the script for keeping score. Currently they just delete on contact. Score is incremented,
but not shown to the player. All levels have their own collectible and they have be added to the scenes. Tutorial
level shows what to collect,  Spring has different color flowers that do nothing, fall will have fake pumpkins that
do damage.