Added fall boss attack, no longer sticking to walls on fall and winter.
Fall boss now throws a homing pumpkin grenade. Grenade updates its
position every 3 seconds and also a new one spawns at the same time. If
it collides with player or hits where the player was when fired, it will
explode. Walls on winter and fall have been fixed. Moving fall platforms
and the area to gather collectibles have not been applied this fix.